requirements.txt
streamlit==1.29.0
tensorflow==2.16.1
numpy==1.24.3
Pillow==10.2.0
matplotlib==3.7.1